social work

+++ src/main/java/com/soapboxrace/core/bo/PresenceBO.java
@@ -18,9 +18,9 @@ public class PresenceBO {
         this.pubSubConnection = this.redisBO.createPubSub();
     }
 
-    public void updatePresence(Long personaId, Integer presence) {
+    public void updatePresence(Long personaId, Long presence) {
         if (personaId != 0L) {
-            Integer currentPresence = this.getPresence(personaId);
+            Long currentPresence = this.getPresence(personaId);
 
             if (!currentPresence.equals(presence)) {
                 this.redisBO.getConnection().sync().set(getPresenceKey(personaId), presence.toString());
@@ -31,18 +31,18 @@ public class PresenceBO {
 
     public void removePresence(Long personaId) {
         if (personaId != 0L) {
-            updatePresence(personaId, 0);
+            updatePresence(personaId, 0L);
             this.redisBO.getConnection().sync().del(getPresenceKey(personaId));
         }
     }
 
-    public int getPresence(Long personaId) {
+    public Long getPresence(Long personaId) {
         String value = this.redisBO.getConnection().sync().get(getPresenceKey(personaId));
 
         if (value == null || value.trim().isEmpty())
-            return 0;
+            return 0L;
 
-        return Integer.parseInt(value);
+        return Long.parseLong(value);
     }
 
     private String getPresenceKey(Long personaId) {
